+++ lib/core/pragma_handler_map.cpp
@@ -36,6 +36,7 @@
 #include <metashell/pragma_pp.hpp>
 #include <metashell/pragma_echo.hpp>
 #include <metashell/pragma_macros.hpp>
+#include <metashell/pragma_macro_names.hpp>
 
 #include <cassert>
 #include <sstream>
@@ -255,6 +256,7 @@ pragma_handler_map pragma_handler_map::build_default(
       .add("metaprogram", "mode", shell_mode("metaprogram", "", false, shell_))
       .add("echo", pragma_echo())
       .add("macros", pragma_macros(shell_))
+      .add("macro", "names", pragma_macro_names(shell_))
       .add("quit", pragma_quit(shell_))
     ;
 }