+++ docs/about/changelog.md
@@ -22,6 +22,7 @@
       (`#msh metaprogram mode`)
     * Add command to display a comment in the shell (`#msh echo`)
     * Add command to display macro definitions (`#msh macros`)
+    * Add command to display macro names (`#msh macro names`)
 
 * Changes to existing behaviour
     * File sections displayed by MDB now shows the source location of the